pyproject.toml
[project]
name = "deepsearchagent-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
]
